Return empty promise when clear database method is called before initialization
Can cause crash without.

+++ js/App/Lib/LocalStorage.js
@@ -320,6 +320,9 @@ export default class TelldusLocalStorage {
 			tableName = 'SensorHistory';
 			uniqueParam = 'sensorId';
 		}
+		if (!db) {
+			return Promise.resolve(null);
+		}
 		return db.executeSql(`DELETE from ${tableName} WHERE ${id} = ${uniqueParam}`).then((res: any): any => {
 			return res;
 		}).catch((error: Object) => {
